benchmarks/pyproject.toml
[project]
authors = [{name = "InfinitiBit Team"}]
dependencies = [
  "graphbit>=0.6.0",
  "numpy>=1.26.2,<2.0.0",
  "matplotlib>=3.9.0",
  "seaborn>=0.13.2",
  "psutil>=7.2.1",
  "autogen-agentchat>=0.4.0",
  "autogen-ext[openai,anthropic,ollama]>=0.4.0",
  "crewai>=1.6.0; python_version < '3.14'",
  "langchain>=1.2.0",
  "langchain-anthropic>=1.3.0",
  "langchain-core>=1.2.6",
  "langchain-ollama>=1.0.1",
  "langchain-openai>=1.1.6",
  "langgraph>=0.2.0",
  "llama-index>=0.14.12",
  "llama-index-llms-anthropic>=0.10.5",
  "llama-index-llms-ollama>=0.9.1",
  "llama-index-llms-azure-openai>=0.3.0",
  "pydantic-ai>=1.22.0",
  "openai>=1.83.0",
  "click>=8.1.0",
  "datasets>=2.14.0",
  "python-dotenv>=1.0.0"
]
description = "Benchmarking package for GraphBit framework"
name = "graphbit-benchmarks"
readme = "README.md"
requires-python = ">=3.10,<3.14"
version = "0.1.0"

[tool.poetry]
authors = ["InfinitiBit Team"]
description = "Benchmarking package for GraphBit framework"
name = "graphbit-benchmarks"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
autogen-agentchat = ">=0.4.0"
autogen-ext = {version = ">=0.4.0", extras = ["openai", "anthropic", "ollama"]}
crewai = {version = ">=1.6.0", python = ">=3.10,<3.14"}
graphbit = ">=0.6.0"
langchain = {version = ">=1.2.0", python = ">=3.10,<4.0"}
langchain-anthropic = {version = ">=1.3.0", python = ">=3.10,<4.0"}
langchain-core = {version = ">=1.2.6", python = ">=3.10,<4.0"}
langchain-ollama = {version = ">=1.0.1", python = ">=3.10,<4.0"}
langchain-openai = {version = ">=1.1.6", python = ">=3.10,<4.0"}
llama-index = ">=0.14.12"
llama-index-llms-anthropic = ">=0.10.5"
llama-index-llms-ollama = ">=0.9.1"
matplotlib = ">=3.9.0"
# Core dependencies with pinned versions for Python 3.10–3.13
numpy = ">=1.26.2,<2.0.0"  # compatible with Python 3.10–3.13
openai = ">=1.83.0"
psutil = ">=7.2.1"
pydantic-ai = {version = ">=1.22.0", python = ">=3.10"}
python = ">=3.10,<3.14"
seaborn = ">=0.13.2"
click = ">=8.1.0"
datasets = ">=2.14.0"
langgraph = {version = ">=0.2.0", python = ">=3.10,<4.0"}
